+++ site/src/components/Resources/AppLink/AppLink.tsx
@@ -6,8 +6,8 @@ import ErrorOutlineIcon from "@mui/icons-material/ErrorOutline"
 import { PrimaryAgentButton } from "components/Resources/AgentButton"
 import { FC } from "react"
 import { combineClasses } from "utils/combineClasses"
-import * as TypesGen from "../../api/typesGenerated"
-import { generateRandomString } from "../../utils/random"
+import * as TypesGen from "../../../api/typesGenerated"
+import { generateRandomString } from "../../../utils/random"
 import { BaseIcon } from "./BaseIcon"
 import { ShareIcon } from "./ShareIcon"
 import { useProxy } from "contexts/ProxyContext"
